+++ errors/bad_syntax_error.py
@@ -1,0 +1,7 @@
+from errors.error import Error
+from position import Interval
+
+
+class BadSyntaxError(Error):
+    def __init__(self, msg: str, interval: Interval) -> None:
+        super().__init__("Bad Syntax", msg, interval)
