+++ errors/unexpected_token_error.py
@@ -1,0 +1,7 @@
+from errors.error import Error
+from position import Interval
+
+
+class UnexpectedTokenError(Error):
+    def __init__(self, msg: str, interval: Interval) -> None:
+        super().__init__("Unexpected Token", msg, interval)
